requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
arrow==1.2.1
asgiref==3.4.1; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.11.1; python_version >= '3.7'
automat==20.2.0
backports.zoneinfo==0.2.1; python_version < '3.9'
boto3==1.20.37
botocore==1.23.37
brotli==1.0.9
cffi==1.15.0
channels-redis==3.3.1
channels==3.0.4
click==8.0.3; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
constantly==15.1.0
coverage==6.2
cryptography==36.0.1; python_version >= '3.6'
daphne==3.0.2; python_version >= '3.6'
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash==2.0.0
deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-autocomplete-light==3.8.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-redis==5.2.0
django-storages==1.12.3
django-widget-tweaks==1.4.12
django==4.0.1
djangorestframework==3.13.1
flask-compress==1.10.1
flask==2.0.2; python_version >= '3.6'
gunicorn==20.1.0
hiredis==2.0.0; python_version >= '3.6'
hyperlink==21.0.0
idna==3.3; python_version >= '3.4'
incremental==21.3.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.1; python_version >= '3.6'
msgpack==1.0.3
numpy==1.19.3
packaging==21.3; python_version >= '3.6'
plotly==5.5.0
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyopenssl==21.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pyparsing==3.0.6; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.1
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
tenacity==8.0.1; python_version >= '3.6'
twisted-iocpsupport==1.0.2; platform_system == 'Windows'
twisted[tls]==21.7.0; python_full_version >= '3.6.7'
txaio==21.2.1; python_version >= '3.6'
typing-extensions==4.0.1; python_version >= '3.6'
tzdata==2021.5; sys_platform == 'win32'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
wheel==0.37.1
wrapt==1.13.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
